feat: take the location-hint on SDK init
