+++ packages/client/src/coordinator/connection/client.ts
@@ -34,6 +34,7 @@ import {
 } from './types';
 import { InsightMetrics, postInsights } from './insights';
 import { version } from '../../../version';
+import { getLocationHint } from './location';
 
 export class StreamClient {
   _user?: UserWithId;
@@ -47,6 +48,8 @@ export class StreamClient {
   key: string;
   listeners: Record<string, Array<(event: StreamVideoEvent) => void>>;
   logger: Logger;
+
+  private locationHint: Promise<string> | undefined;
 
   node: boolean;
   options: StreamClientOptions;
@@ -100,6 +103,10 @@ export class StreamClient {
 
     this.browser = inputOptions.browser || typeof window !== 'undefined';
     this.node = !this.browser;
+
+    if (this.browser) {
+      this.locationHint = getLocationHint(options?.locationHintTimeout);
+    }
 
     this.options = {
       timeout: 5000,
@@ -176,6 +183,17 @@ export class StreamClient {
     this.wsBaseURL = this.baseURL
       .replace('http', 'ws')
       .replace(':3030', ':8800');
+  };
+
+  getLocationHint = async (timeout?: number): Promise<string> => {
+    const hint = await this.locationHint;
+    if (!hint || hint === 'ERR') {
+      this.locationHint = getLocationHint(
+        timeout ?? this.options.locationHintTimeout,
+      );
+      return this.locationHint;
+    }
+    return hint;
   };
 
   _getConnectionID = () =>
