+++ packages/client/src/coordinator/connection/types.ts
@@ -103,6 +103,10 @@ export type StreamClientOptions = Partial<AxiosRequestConfig> & {
   logger?: Logger;
   logLevel?: LogLevel;
   /**
+   * The default timeout for requesting a location hint.
+   */
+  locationHintTimeout?: number;
+  /**
    * When true, user will be persisted on client. Otherwise if `connectUser` call fails, then you need to
    * call `connectUser` again to retry.
    * This is mainly useful for chat application working in offline mode, where you will need client.user to
